pyproject.toml
[project]
name = "vector-domain-search-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrew-GLB",email = "andrexcobar@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlmodel (>=0.0.31,<0.0.32)",
    "typesense (>=1.3.0,<2.0.0)",
    "polars (>=1.37.1,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pypdf (>=6.6.0,<7.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-genai (>=1.60.0,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "adbc-driver-postgresql (>=1.10.0,<2.0.0)",
    "pandas (>=3.0.0,<4.0.0)",
    "pyarrow (>=23.0.0,<24.0.0)",
    "connectorx (>=0.4.5,<0.5.0)",
    "fpdf2 (>=2.8.5,<3.0.0)",
    "ruff (>=0.14.14,<0.15.0)",
    "faker (>=40.1.2,<41.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mypy (>=1.19.1,<2.0.0)"
]

# --- Mypy Configuration (Requirement #26) ---
[tool.mypy]
# Uses the plugin bundled with Pydantic V2
plugins = ["pydantic.mypy"]

python_version = "3.12"
strict = true
ignore_missing_imports = true
explicit_package_bases = true

# This ensures mypy understands your 'app.' imports correctly
mypy_path = "."

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = false
disallow_subclassing_any = false
disallow_untyped_decorators = false

[tool.ruff]
# Target Python 3.12 features (e.g., f-string improvements, type parameter syntax)
target-version = "py312"
line-length = 88

[tool.ruff.lint]
# E, W: pycodestyle (PEP-8)
# F: Pyflakes (Logic errors)
# I: isort (Import sorting)
# D: pydocstyle (Docstrings - your specific requirement)
# UP: pyupgrade (Modernize Python syntax)
# B: flake8-bugbear (Common design flaws)
# ANN: flake8-annotations (Enforce type hints for FastAPI/OpenAPI)
# ASYNC: flake8-async (Async best practices)
# FAST: FastAPI specific rules
# RUF: Ruff-specific rules
select = ["E", "W", "F", "I", "D", "UP", "B", "ANN", "ASYNC", "FAST", "RUF"]

# Specific ignores to keep your workflow smooth
ignore = [
    "D100",    # Missing docstring in public module (too noisy for small files)
    "D101",    # Missing docstring in public class
    "D102",    # Missing docstring in public method
    "D104",    # Missing docstring in public package (__init__.py)
    "ANN101",  # Missing type annotation for 'self' (deprecated in Ruff)
    "ANN102",  # Missing type annotation for 'cls'
    "D107",    # Missing docstring in __init__ (class docstring is usually enough)
    "B008",    # Do not perform function calls in argument defaults (Needed for FastAPI Depends)
    "E501",
    "D205",
    "B904",
    "W293",
    "ANN401",
    "RUF001"
]

[tool.ruff.lint.pydocstyle]
# Google is the cleanest convention for FastAPI/Pydantic projects
convention = "google"

[tool.ruff.lint.per-file-ignores]
# Tests don't need docstrings and can use 'assert'
"tests/*" = ["D", "S101"]
# Migrations usually have long lines and don't need docs
"migrations/*" = ["E501", "D"]

[tool.ruff.format]
# Keep these to handle your styling automatically
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
# NEW in 2024/2025: Formats code snippets inside your docstrings!
docstring-code-format = true

[tool.ruff.lint.isort]
# Groups imports logically
combine-as-imports = true
lines-after-imports = 2